Move `SafeFireAndForgotExtensions`
