+++ src/CommunityToolkit.Maui/Extensions/SafeFireAndForgotExtensions.shared.cs
@@ -1,4 +1,4 @@
-namespace CommunityToolkit.Maui.Behaviors;
+namespace CommunityToolkit.Maui.Extensions;
 
 static class SafeFireAndForgetExtensions
 {
